Make copyright disclaimer more visible and expand on permission of use
In some cases relinquishing copyright alone is not sufficient to grant
permission of use, such as in jurisdictions that do not recognize
public domain grants or when other kinds of IP claims would restrict
this.
This makes those permissions explicit using the minimal language from
ISC/0BSD for any contexts where they are not otherwise granted by the
forfeit of copyright.